separate business and app js and modularize
