+++ templates/package.json
@@ -1,34 +1,34 @@
 {
   "dependencies": {
     "archieml": "^0.3.1",
-    "babel-core": "^6.4.0",
-    "babel-loader": "^6.2.1",
-    "babel-preset-es2015": "^6.3.13",
-    "browser-sync": "^2.11.0",
-    "cheerio": "^0.19.0",
+    "babel-core": "^6.5.1",
+    "babel-loader": "^6.2.2",
+    "babel-preset-es2015": "^6.5.0",
+    "browser-sync": "^2.11.1",
+    "cheerio": "^0.20.0",
     "del": "^2.2.0",
-    "gulp": "^3.9.0",
+    "gulp": "^3.9.1",
     "gulp-autoprefixer": "^3.1.0",
-    "gulp-babel": "^6.1.1",
     "gulp-callback": "0.0.3",
     "gulp-changed": "^1.3.0",
     "gulp-file-include": "^0.13.7",
     "gulp-hb": "^2.6.5",
     "gulp-htmlmin": "^1.3.0",
     "gulp-notify": "^2.2.0",
-    "gulp-plumber": "^1.0.1",
+    "gulp-plumber": "^1.1.0",
     "gulp-rename": "^1.2.2",
     "gulp-replace": "^0.5.4",
     "gulp-smoosher": "0.0.9",
-    "gulp-stylus": "^2.2.0",
+    "gulp-stylus": "^2.3.0",
     "gulp-util": "^3.0.7",
-    "gulp-zip": "^3.0.2",
-    "request": "^2.67.0",
+    "gulp-zip": "^3.1.0",
+    "json-loader": "^0.5.4",
+    "request": "^2.69.0",
     "require-dir": "^0.3.0",
     "run-sequence": "^1.1.5",
-    "shelljs": "^0.5.3",
-    "underscore.string": "^3.2.2",
+    "shelljs": "^0.6.0",
+    "underscore.string": "^3.2.3",
     "webpack-stream": "^3.1.0",
-    "yargs": "^3.31.0"
+    "yargs": "^3.32.0"
   }
 }
