+++ templates/src/html/partials/base/base-critical.hbs
@@ -1,4 +1,5 @@
-<script type='text/javascript'>
+<script>
+	// TODO make module
 	/*! loadJS: load a JS file asynchronously. [c]2014 @scottjehl, Filament Group, Inc. (Based on http://goo.gl/REQGQ by Paul Irish). Licensed MIT */
 	(function( w ){
 		var loadJS = function( src, cb ){
@@ -122,3 +123,7 @@
 	  loadCSS( fontFileUrl );
 	}( this ));
 </script>
+
+
+
+<script src='js/business.js' async=true></script>
