--- templates/src/html/partials/base/base-js-webpack.hbs
@@ -1,7 +1,0 @@
-
-<!-- htmlmin:ignore -->
-<!-- smoosh -->
-<script src='js/bundle.js' type='text/javascript'></script>
-<!-- endsmoosh -->
-<!-- htmlmin:ignore -->
-
