+++ templates/src/html/partials/base/base-js.hbs
@@ -1,7 +1,6 @@
-<!-- js libs -->
-<script src='https://apps.bostonglobe.com/common/js/jquery/jquery-1.11.2.min.js' type='text/javascript'></script>
-
+<!-- htmlmin:ignore -->
 <!-- smoosh -->
-<script src='js/util.js' type='text/javascript'></script>
-<script src='js/main.js' type='text/javascript'></script>
+<script src='js/bundle.js' type='text/javascript'></script>
 <!-- endsmoosh -->
+<!-- htmlmin:ignore -->
+
