+++ templates/src/js/app.js
@@ -1,0 +1,3 @@
+import setupDefaultApps from './utils/setupDefaultApps';
+
+setupDefaultApps();
