--- templates/src/js/main-webpack.js
@@ -1,3 +1,0 @@
-import setupDefaultApps from './utils/setupDefaultApps';
-
-setupDefaultApps();
