--- templates/src/js/utils/disableScroll.js
@@ -1,9 +1,0 @@
-export default function disableScroll() {
-	if (window.addEventListener) {
-		window.addEventListener('DOMMouseScroll', preventDefault, false);
-	}
-	window.onwheel = preventDefault; // modern standard
-	window.onmousewheel = document.onmousewheel = preventDefault; // older browsers, IE
-	window.ontouchmove  = preventDefault; // mobile
-	document.onkeydown  = preventDefaultForScrollKeys;
-};
