--- templates/src/js/utils/enableScroll.js
@@ -1,9 +1,0 @@
-export default function enableScroll() {
-	if (window.removeEventListener) {
-		window.removeEventListener('DOMMouseScroll', preventDefault, false);
-	}
-	window.onmousewheel = document.onmousewheel = null;
-	window.onwheel = null;
-	window.ontouchmove = null;
-	document.onkeydown = null;
-};
