+++ templates/src/js/utils/misc/loadjs.js
@@ -1,0 +1,14 @@
+/*! loadJS: load a JS file asynchronously. [c]2014 @scottjehl, Filament Group, Inc. (Based on http://goo.gl/REQGQ by Paul Irish). Licensed MIT */
+
+export default function loadJS( src, cb ) {
+	"use strict";
+	var ref = document.getElementsByTagName( "script" )[ 0 ];
+	var script = document.createElement( "script" );
+	script.src = src;
+	script.async = true;
+	ref.parentNode.insertBefore( script, ref );
+	if (cb && typeof(cb) === "function") {
+		script.onload = cb;
+	}
+	return script;
+}
