+++ templates/src/js/utils/paywall/disableScroll.js
@@ -1,0 +1,12 @@
+import preventDefault from './preventDefault';
+import preventDefaultForScrollKeys from './preventDefaultForScrollKeys';
+
+export default function disableScroll() {
+	if (window.addEventListener) {
+		window.addEventListener('DOMMouseScroll', preventDefault, false);
+	}
+	window.onwheel = preventDefault; // modern standard
+	window.onmousewheel = document.onmousewheel = preventDefault; // older browsers, IE
+	window.ontouchmove  = preventDefault; // mobile
+	document.onkeydown  = preventDefaultForScrollKeys;
+};
