--- templates/src/js/utils/preventDefaultForScrollKeys.js
@@ -1,7 +1,0 @@
-export default function preventDefaultForScrollKeys(e) {
-	var keys = {38: 1, 40: 1, 32: 1};
-	if (keys[e.keyCode]) {
-		preventDefault(e);
-		return false;
-	}
-};
